requirements.txt
fastapi
uvicorn
pandas
transformers
datasets
sentence-transformers
faiss-cpu
bert-score
